Refactor sortable field GraphQL enums to top-level classes
