+++ src/main/java/com/demo/filmdb/utils/SortUtil.java
@@ -14,54 +14,6 @@ import java.util.stream.Collectors;
 public class SortUtil {
 
     /**
-     * Represents sortable fields of a film. Must match GraphQL enum with the same name.
-     */
-    public enum SortableFilmField {
-        ID("id"),
-        TITLE("title"),
-        RELEASE_DATE("releaseDate");
-
-        private final String fieldName;
-
-        SortableFilmField(String fieldName) {
-            this.fieldName = fieldName;
-        }
-
-        /**
-         * Returns corresponding entity field name
-         *
-         * @return field name
-         */
-        public String getFieldName() {
-            return fieldName;
-        }
-    }
-
-    /**
-     * Represents sortable fields of a person. Must match GraphQL enum with the same name.
-     */
-    public enum SortablePersonField {
-        ID("id"),
-        NAME("name"),
-        DATE_OF_BIRTH("dateOfBirth");
-
-        private final String fieldName;
-
-        SortablePersonField(String fieldName) {
-            this.fieldName = fieldName;
-        }
-
-        /**
-         * Returns corresponding entity field name
-         *
-         * @return field name
-         */
-        public String getFieldName() {
-            return fieldName;
-        }
-    }
-
-    /**
      * Filters out {@code pageable}'s {@link Sort} properties that aren't annotated with @{@link Sortable} in the
      * {@code entityClass}.
      * @param pageable containing the {@link Sort} to filter.
